changeUser3

+++ src/main/resources/db/migration/V1__Create_user_table.sql
@@ -1,6 +1,6 @@
 CREATE table USER
 (
-    id int AUTO_INCREMENT PRIMARY KEY NOT NULL,
+    id bigint AUTO_INCREMENT PRIMARY KEY NOT NULL,
     account_id VARCHAR (100),
     name VARCHAR (50),
     token VARCHAR (36),
